employee-by-company/hello_world/requirements.txt
requests
pandas
snowflake-connector-python